+++ metadata.rb
@@ -13,15 +13,15 @@
 # limitations under the License.
 
 name             'vagrant'
-maintainer       'Joshua Timberman'
+maintainer       'Sous-chefs'
 maintainer_email 'cookbooks@housepub.org'
 license          'Apache-2.0'
 description      'Installs Vagrant and provides a vagrant_plugin LWRP for installing Vagrant plugins.'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.7.2'
 
-source_url       'https://github.com/jtimberman/vagrant-cookbook' if respond_to?(:source_url)
-issues_url       'https://github.com/jtimberman/vagrant-cookbook/issues' if respond_to?(:issues_url)
+source_url       'https://github.com/sous-chefs/vagrant-cookbook' if respond_to?(:source_url)
+issues_url       'https://github.com/sous-chefs/vagrant-cookbook/issues' if respond_to?(:issues_url)
 
 supports         'debian', '>= 6.0'
 supports         'ubuntu', '>= 12.04'